server/project/requirements.txt
asgiref==3.11.0
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.9
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==25.0
pillow==12.0.0
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg2-binary==2.9.11
qrcode==8.2
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.2
whitenoise==6.11.0
